// signal-examples/src/app/models/employee.ts
export interface EmployeeModel {
    firstname: string;
    lastname: string;
    email: string;
    phoneno: string;
    address: string;
    iscontract: boolean;
}

export interface EmployeeListModel {
    id: string;
    firstname: string;
    lastname: string;
    email: string;
    phoneno: string;
    address: string;
    iscontract: boolean;
}